refactor: collapse OpenAI-compatible providers and search actions
- createOpenAICompatibleProvider: one class parameterized by
  {key, name, configFields, parseAndValidate, resolve,
  listDefaultModels, supportsEmbeddings}; openai, anthropic, gemini,
  groq, lmstudio and lemonade are now spec objects, and the 8
  five-line alias LLM/embedding classes are gone
- createSearchAction: web/academic/social search actions share one
  execute body and schema, differing only in prompts, enabled
  predicate and engine list; drops the vestigial type literal from
  the web_search schema
- extractor prompt deduplicated into prompts/search/extractor.ts;
  the scrapeURL copy carried 18 spaces of indentation per line paid
  on every extraction call

Net: -428 lines, no behavior change.
