+++ src/lib/agents/search/researcher/actions/search/socialSearch.ts
@@ -1,11 +1,4 @@
-import z from 'zod';
-import { ResearchAction } from '../../../types';
-import { ResearchBlock } from '@/lib/types';
-import { executeSearch } from './baseSearch';
-
-const schema = z.object({
-  queries: z.array(z.string()).describe('List of social search queries'),
-});
+import { createSearchAction } from './createSearchAction';
 
 const socialSearchDescription = `
 Use this tool to perform social media searches for relevant posts, discussions, and trends related to the user's query. Provide a list of concise search queries that will help gather comprehensive social media information on the topic at hand.
@@ -19,43 +12,15 @@ For example, if the user is interested in public opinion on electric vehicles, y
 If this tool is present and no other tools are more relevant, you MUST use this tool to get the needed social media information.
 `;
 
-const socialSearchAction: ResearchAction<typeof schema> = {
+const socialSearchAction = createSearchAction({
   name: 'social_search',
-  schema: schema,
-  getDescription: () => socialSearchDescription,
-  getToolDescription: () =>
+  toolDescription:
     "Use this tool to perform social media searches for relevant posts, discussions, and trends related to the user's query. Provide a list of concise search queries that will help gather comprehensive social media information on the topic at hand.",
+  getDescription: () => socialSearchDescription,
   enabled: (config) =>
     config.sources.includes('discussions') &&
     config.classification.classification.discussionSearch === true,
-  execute: async (input, additionalConfig) => {
-    input.queries = (
-      Array.isArray(input.queries) ? input.queries : [input.queries]
-    ).slice(0, 3);
-
-    const researchBlock = additionalConfig.session.getBlock(
-      additionalConfig.researchBlockId,
-    ) as ResearchBlock | undefined;
-
-    if (!researchBlock) throw new Error('Failed to retrieve research block');
-
-    const results = await executeSearch({
-      llm: additionalConfig.llm,
-      embedding: additionalConfig.embedding,
-      mode: additionalConfig.mode,
-      queries: input.queries,
-      researchBlock: researchBlock,
-      session: additionalConfig.session,
-      searchConfig: {
-        engines: ['reddit'],
-      },
-    });
-
-    return {
-      type: 'search_results',
-      results: results,
-    };
-  },
-};
+  engines: ['reddit'],
+});
 
 export default socialSearchAction;
