--- src/lib/models/providers/anthropic/anthropicLLM.ts
@@ -1,5 +1,0 @@
-import OpenAILLM from '../openai/openaiLLM';
-
-class AnthropicLLM extends OpenAILLM {}
-
-export default AnthropicLLM;
