--- src/lib/models/providers/gemini/geminiEmbedding.ts
@@ -1,5 +1,0 @@
-import OpenAIEmbedding from '../openai/openaiEmbedding';
-
-class GeminiEmbedding extends OpenAIEmbedding {}
-
-export default GeminiEmbedding;
