--- src/lib/models/providers/gemini/geminiLLM.ts
@@ -1,5 +1,0 @@
-import OpenAILLM from '../openai/openaiLLM';
-
-class GeminiLLM extends OpenAILLM {}
-
-export default GeminiLLM;
