--- src/lib/models/providers/groq/groqLLM.ts
@@ -1,5 +1,0 @@
-import OpenAILLM from '../openai/openaiLLM';
-
-class GroqLLM extends OpenAILLM {}
-
-export default GroqLLM;
