--- src/lib/models/providers/lemonade/lemonadeEmbedding.ts
@@ -1,5 +1,0 @@
-import OpenAIEmbedding from '../openai/openaiEmbedding';
-
-class LemonadeEmbedding extends OpenAIEmbedding {}
-
-export default LemonadeEmbedding;
