--- src/lib/models/providers/lemonade/lemonadeLLM.ts
@@ -1,5 +1,0 @@
-import OpenAILLM from '../openai/openaiLLM';
-
-class LemonadeLLM extends OpenAILLM {}
-
-export default LemonadeLLM;
