--- src/lib/models/providers/lmstudio/lmstudioEmbedding.ts
@@ -1,5 +1,0 @@
-import OpenAIEmbedding from '../openai/openaiEmbedding';
-
-class LMStudioEmbedding extends OpenAIEmbedding {}
-
-export default LMStudioEmbedding;
