--- src/lib/models/providers/lmstudio/lmstudioLLM.ts
@@ -1,5 +1,0 @@
-import OpenAILLM from '../openai/openaiLLM';
-
-class LMStudioLLM extends OpenAILLM {}
-
-export default LMStudioLLM;
